Handle completed function calls in streams (#10)
* make `SSEStreamHandler` generic

* implement `runRequiresActionKey`

* add helper `handleUnknownEvent`

* Implement function response handling

* add streaming capability to `SubmitToolOutputs`

* handle `runRequiresAction` (Function Calling)

* Documentation

* formatting

* remove invalid key

* Update Sources/GPTBridge/Models/RunHandlers/Interface/FunctionRunHandlable.swift


* safely unwrap url


* Update Sources/GPTBridge/Internal/SSE Handling/StreamingRequestManager.swift


* Update Sources/GPTBridge/Internal/SSE Handling/StreamingRequestManager.swift


* unwrap `StreamingRequestManager`

* correct return type

* convert repeated guard into throwing init

* remove unused Error enums

---------
