+++ Sources/GPTBridge/Models/RunHandlers/FunctionRunHandler.swift
@@ -7,19 +7,27 @@
 
 import Foundation
 
-class FunctionRunHandler: FunctionRunHandlable {
-    var functionParameters: [String : FunctionArgument]?
-
-    var requiredAction: RequiredAction?
-    
-    var runThreadResponse: RunThreadResponse
-    
+open class FunctionRunHandler: FunctionRunHandlable {
+    public private(set) var functionParameters: [String : FunctionArgument]?
+    private(set) var requiredAction: RequiredAction?
+    var runThreadResponse: RunThreadResponse?
+    /// set `functionParameters` when `runThreadResponse is non-nil `using `parse`
     func handle() async throws {
         self.functionParameters = try parse()
     }
 
-    init(runThreadResponse: RunThreadResponse) {
+    convenience init(functionParameters: [String: FunctionArgument]) {
+        self.init(runThreadResponse: nil)
+        self.functionParameters = functionParameters
+    }
+
+    init(runThreadResponse: RunThreadResponse?) {
         self.runThreadResponse = runThreadResponse
-        self.requiredAction = runThreadResponse.requiredAction
+        self.requiredAction = runThreadResponse?.requiredAction
+        if runThreadResponse != nil {
+            Task {
+                try await self.handle()
+            }
+        }
     }
 }
