+++ Sources/GPTBridge/Models/RunHandlers/Interface/FunctionRunHandlable.swift
@@ -17,13 +17,17 @@ extension FunctionRunHandlable {
     private var noActionRetrievedError: NSError {
         NSError(domain: "noActionRetrievedError", code: 0, userInfo: nil)
     }
-
+    /// Parse `RequiredAction` to retrieve function calls (`ToolCall`)
+    /// - throws: `noActionRetrievedError` if `requiredAction` is nil or there are no function calls
+    /// - returns: [ToolCall.id: FunctionArgument]
     func parse() throws -> [String: FunctionArgument] {
-       guard let action = requiredAction,
-             action.submitToolOutputs.toolCalls.count > 0
-       else {
-           throw noActionRetrievedError
-       }
-       return action.submitToolOutputs.toolCalls[0].function.arguments // TODO: Generic arguments
+        guard let action = requiredAction,
+              action.submitToolOutputs.toolCalls.count > 0
+        else {
+            throw noActionRetrievedError
+        }
+        return action.submitToolOutputs.toolCalls.reduce(into: [:]) { partialResult, toolCall in
+            partialResult[toolCall.id] = toolCall.function.arguments.first?.value
+        }
     }
 }
