+++ Sources/GPTBridge/Models/RunHandlers/Interface/RunHandler.swift
@@ -9,7 +9,7 @@ import Foundation
 
 protocol RunHandler {
     /// The run's response, decoded
-    var runThreadResponse: RunThreadResponse { get }
+    var runThreadResponse: RunThreadResponse? { get }
     /// Handle the run - implemented differently for different handlers
     func handle() async throws
 }
